+++ lib/models/expense.dart
@@ -1,0 +1,60 @@
+class Expense {
+  final String id;
+  final double amount;
+  final String category;
+  final String description;
+  final DateTime date;
+  final DateTime createdAt;
+
+  Expense({
+    required this.id,
+    required this.amount,
+    required this.category,
+    required this.description,
+    required this.date,
+    required this.createdAt,
+  });
+
+  // Convert Expense to JSON
+  Map<String, dynamic> toJson() {
+    return {
+      'id': id,
+      'amount': amount,
+      'category': category,
+      'description': description,
+      'date': date.toIso8601String(),
+      'createdAt': createdAt.toIso8601String(),
+    };
+  }
+
+  // Create Expense from JSON
+  factory Expense.fromJson(Map<String, dynamic> json) {
+    return Expense(
+      id: json['id'] as String,
+      amount: (json['amount'] as num).toDouble(),
+      category: json['category'] as String,
+      description: json['description'] as String,
+      date: DateTime.parse(json['date'] as String),
+      createdAt: DateTime.parse(json['createdAt'] as String),
+    );
+  }
+
+  // Copy with method for updates
+  Expense copyWith({
+    String? id,
+    double? amount,
+    String? category,
+    String? description,
+    DateTime? date,
+    DateTime? createdAt,
+  }) {
+    return Expense(
+      id: id ?? this.id,
+      amount: amount ?? this.amount,
+      category: category ?? this.category,
+      description: description ?? this.description,
+      date: date ?? this.date,
+      createdAt: createdAt ?? this.createdAt,
+    );
+  }
+}
