--- lib/screen/start_screen.dart
@@ -1,30 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:heal_tether_task/screen/home_screen.dart';
-
-class StartScreen extends StatelessWidget {
-  const StartScreen({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        backgroundColor: Color(0xFF012B5B),
-        automaticallyImplyLeading: false,
-        centerTitle: true,
-        title: const Text('Start Screen'),
-      ),
-      body: Center(
-        child: ElevatedButton(
-          style: ElevatedButton.styleFrom(backgroundColor: Color(0xFF012B5B)),
-          onPressed: () {
-            Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => const HomeScreen()),
-            );
-          },
-          child: const Text('Go to Home Screen'),
-        ),
-      ),
-    );
-  }
-}
