+++ pubspec.yaml
@@ -36,6 +36,7 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
   flutter_launcher_icons: ^0.13.1
+  shared_preferences: ^2.2.2
 
 dev_dependencies:
   flutter_test:
